Flex the right direction
At first I thought this needed to be a flex-col.
Turned out that it was a plain flex (-wrap) situation. It just needed
the magic div aroudn the second and third feature boxes. I also adjusted
the max-w from 2xl to 6xl and made it cahnge on the lg: breakpoint. This
means that it stays as a mobile view until it's large enough for the
desktop view.